pyproject.toml
[project]
name = "ticket-analyzer-api"
version = "0.1.0"
description = "AI-powered backend service that classifies support tickets using LLMs and returns validated structured JSON responses."
authors = [
    {name = "Leonardo Rudolfo Cardoso"}
]
license = {text = "MIT"}
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "openai (>=1.60.0,<2.0.0)",
    "pydantic-settings (>=2.7.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[dependency-groups]
dev = [
    "pyright (>=1.1.408,<2.0.0)",
    "ruff (>=0.15.1,<0.16.0)",
    "pytest (>=9.0.0,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.0,<1.0.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"


[tool.poetry]
packages = [{ include = "ticket_analyzer" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
